+++ mp3splt-project/newmp3splt/slackware/make_slackware_package.sh
@@ -12,29 +12,37 @@ echo
 echo $'Package :\tslackware'
 echo
 
-SLACK_TEMP=/tmp/slack_temp
+#if we don't have the distribution file
+DIST_FILE="../mp3splt-${MP3SPLT_VERSION}-${ARCH}.tgz"
+if [[ ! -f $DIST_FILE ]];then
+    SLACK_TEMP=/tmp/slack_temp
+    
+    #we set the necessary flags
+    #export CFLAGS="-O2 -march=$ARCH -mcpu=$ARCH"
+    export CFLAGS="$CFLAGS"
+    export LDFLAGS="$LDFLAGS"
 
-#we set the necessary flags
-#export CFLAGS="-O2 -march=$ARCH -mcpu=$ARCH"
-export CFLAGS="$CFLAGS"
-export LDFLAGS="$LDFLAGS"
+    #we create the needed directories
+    DATEMV=`date +-%d_%m_%Y__%H_%M_%S`
+    if [[ -e $SLACK_TEMP ]];then
+        mv $SLACK_TEMP $SLACK_TEMP/mp3splt${DATEMV}
+    fi
+    mkdir -p $SLACK_TEMP/mp3splt/usr/doc/mp3splt
+    mkdir -p $SLACK_TEMP/mp3splt/install
 
-#we create the needed directories
-DATEMV=`date +-%d_%m_%Y__%H_%M_%S`
-if [[ -e $SLACK_TEMP ]];then
-    mv $SLACK_TEMP $SLACK_TEMP/mp3splt${DATEMV}
+    #we compile
+    ./autogen.sh &&\
+        ./configure --prefix=/usr &&\
+        make clean &&\
+        make &&\
+        make DESTDIR=$SLACK_TEMP/mp3splt install &&\
+        cp $MP3SPLT_DOC_FILES $SLACK_TEMP/mp3splt/usr/doc/mp3splt &&\
+        cp slackware/slack-* $SLACK_TEMP/mp3splt/install &&\
+        cd $SLACK_TEMP/mp3splt &&\
+        /sbin/makepkg -l y -c y mp3splt-${MP3SPLT_VERSION}-$ARCH.tgz &&\
+        mv mp3splt-${MP3SPLT_VERSION}-$ARCH.tgz $PROGRAM_DIR/.. || exit 1
+else
+    echo
+    echo "We already have the $DIST_FILE distribution file !"
+    echo
 fi
-mkdir -p $SLACK_TEMP/mp3splt/usr/doc/mp3splt
-mkdir -p $SLACK_TEMP/mp3splt/install
-
-#we compile
-./autogen.sh &&\
-./configure --prefix=/usr &&\
-make clean &&\
-make &&\
-make DESTDIR=$SLACK_TEMP/mp3splt install &&\
-cp $MP3SPLT_DOC_FILES $SLACK_TEMP/mp3splt/usr/doc/mp3splt &&\
-cp slackware/slack-* $SLACK_TEMP/mp3splt/install &&\
-cd $SLACK_TEMP/mp3splt &&\
-/sbin/makepkg -l y -c y mp3splt-${MP3SPLT_VERSION}-$ARCH.tgz &&\
-mv mp3splt-${MP3SPLT_VERSION}-$ARCH.tgz $PROGRAM_DIR/.. || exit 1
